requirements.txt
requests
google-genai
groq
slack_sdk
PyYAML
yara-python
docker
supabase
pytz
tenacity